Implement recorded games data tester
Now the port games have almost perfect compatibility with the original SWOS, and the tests can be run in the future to ensure it stays that way. Besides that plenty of small bug fixes, mostly discovered by the tester.

- [ida2asm] fixed setting invalid operand size for instructions that access struct fields, wrongly overriding size specifier (byte/word/dword) with field size (e.g. cmp word ptr ballSprite.z+2, 8)
- [ida2asm] added support for exporting multidimensional arrays
- [ida2asm] now generates proper output when esp is directly referenced
- [ida2asm] added support for label aliases: labels having address as some other symbol (was needed to mark animation and frame indices tables start & end)
- [ida2asm] fixed accessing invalid replacement tokens as they got deallocated from the stack, made them static
- [ida2asm] removed class from Offsets enum, since it's really more like an int, and needs to be available for calculations
- [ida2asm] fixed invalid variable offsets generated when exporting declaration
- [ida2asm] removed 4-byte alignment on large exported structures
- [ida2asm] fixed some text output misalignment in comments
- [ida2asm] removed extraneous space after type name for exported variables with alignment specifier
- [ida2asm] fixed offset generation when real variable after alias variables has an alignment
- [ida2asm] fixed invalid data bytes generated when exported variable is declared as a struct type over an ordinary variable in assembly source
- [ida2asm] fixed alignment directive not working on exported structs
- [ida2asm] made registerPointer() return -1 if the point pool is filled
- fixed menu background disappearing when asset resolution changes
- fixed a texture loading bug which happens when a team's bench texture is already present in the cache and doesn't need to be recreated
- fixed mistake in pitch type generating routine, condition to pick seasonal or random pitch type was inverted
- fixed bench not working when coaching-only a team
- fixed wrong x coordinate of a player that's coming in to the game from the bench
- fixed invalid minimum height expression when setting window size
- fixed the game length, it's now really 3 minutes
- fixed the code to detect if any player is using a particular key and block it from invoking some game functionality (e.g. S for stats)
- fixed FPS going wild occasionally after alt-tab
- fixed selected player menu index not getting reset in bench between matches, and a few more variables
- fixed "local variable" in UpdatePlayersAndBall(), i.e. removed it, it's not really a variable, top of stack is read
- reset all controls static variables between matches
- fixed invalid width in sprite 1198 (player number 11)
- fixed potential source of nondeterminism in bench controls (reset pick bench team counter)
- fixed a few bugs when substituting a player (one that was causing skill values overwrite, didn't set A4 to team pointer before calling ApplyTactics())
- fixed slightly invalid logic when camera was leaving subs menu
- fixed code for detecting if referee reached his destination
- fixed code to call the bench
- stopped unloading menu background when game starts (and loading it back when it ends), unless there's not enough VRAM to load the sprites
- did a small refactoring of the bench and camera code, removed some unnecessary stuff
- added bunch of sprites to the repo that were somehow missed last time
- fixed 32-bit libjpeg-9.dll and libpng16-16.dll being in 64-bit DLL directory
- fixed invalid sprite width in several characters (especially in big font)
- improved upon SWOS pre-match variable initialization
- synchronized bench invoke logic with original SWOS
- checked update bench code and fixed return to game action to react only when clear left/right was pressed, without up/down, in few places
- fixed menu error, it was missing #pragma pack
- switched to FixedPoint's instead of floats for replays
- finished camera documentation
- wrote a bit more ida2asm documentation
- refactored directory traversal into a single routine with callback, made all users use it instead of doing very similar opendir()/readdir() loops
- fixed directory traversing in mock files
- fixed crash reporter error at startup in 64-bit version, complaining it can't find the language file
- added tests for CalculateDeltaXAndY()
- upgraded tests to VS2022
- added SWOS++ and SWPE to the repository

+++ tests/src/mocks/mockTimer.cpp
@@ -2,8 +2,8 @@
 
 void initTimer() {}
 void initFrameTicks() {}
-double targetFps() { return 0; }
-void setTargetFps(double) {}
+int targetFps() { return 0; }
+void setTargetFps(int) {}
 void timerProc(int) {}
 void markFrameStartTime() {}
 void menuFrameDelay() {}
